Included framework for second calculator. Started TNB 35.
