+++ JavaLearning/src/Calc2.java
@@ -1,0 +1,51 @@
+import java.util.Scanner;
+
+public class Calc2 {
+	public static void main(String[] args){
+		switch(){
+		case 0: addition();
+		break;
+		case 1: subtraction();
+		break;
+		case 2: multiplication();
+		break;
+		case 3: division();
+		break;
+		case 4: modulus();
+		break;
+		default: choice();
+		}
+	}
+	//methods for math
+	public static void addition(){
+		double addArray[] = {};
+	}
+	public static void subtraction(){
+		double subArray[] = {};
+	}
+	public static void multiplication(){
+		double multArray[] = {};
+	}
+	public static void division(){
+		double divArray[] = {};
+	}
+	public static void modulus(){
+		double modArray[] = {};
+	}
+	//Methods for User Input
+	public static void input(){
+		Scanner userInput = new Scanner(System.in);
+	}
+	public static void choice(){
+		System.out.println("Please select an operator: ");
+		if(name != String){
+			System.out.println("Please select an option!");
+		}else{
+			System.out.println("Thanks for your selection!");
+		}
+	}
+	//Methods for errors
+	public void errors(){
+		
+		}
+	}
