+++ JavaLearning/src/Learning.java
@@ -1,51 +1,15 @@
-import java.util.Scanner;
 
 public class Learning {
 	public static void main(String[] args){
-		switch(){
-		case 0: addition();
-		break;
-		case 1: subtraction();
-		break;
-		case 2: multiplication();
-		break;
-		case 3: division();
-		break;
-		case 4: modulus();
-		break;
-		default: choice();
+		System.out.println(average(43, 56, 73, 8, 65, 76, 2, 31));
+	}
+	
+	public static int average(int...numbers){
+		int total = 0;
+		for(int x:numbers){
+			total += x;
 		}
+		return total/numbers.length;
 	}
-	//methods for math
-	public static void addition(){
-		double addArray[] = {};
-	}
-	public static void subtraction(){
-		double subArray[] = {};
-	}
-	public static void multiplication(){
-		double multArray[] = {};
-	}
-	public static void division(){
-		double divArray[] = {};
-	}
-	public static void modulus(){
-		double modArray[] = {};
-	}
-	//Methods for User Input
-	public static void input(){
-		Scanner userInput = new Scanner(System.in);
-	}
-	public static void choice(){
-		System.out.println("Please select an operator: ");
-		if(name != String){
-			System.out.println("Please select an option!");
-		}else{
-			System.out.println("Thanks for your selection!");
-		}
-	}
-	//Methods for errors
-	public void errors(){
-		
-		}
-	}
+	
+}
